Homepage full responsive again and structure configurations.
I have re-designed the Homepage to be responsive in all resolutions. On the other hand the App.js structure has been changed to a unique <Router> section. Now the Body.js no longer exists.

PD: In this work session i had a issue with babel and jsx, appearing an error when i established ShowFlashMessages.js.
The error was an import like: "/users/gabriel/desktop... etc. ShowFlashMessages.js".
It cannot be like that.